requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
qdrant-client==1.9.1
redis==5.0.7
psycopg[binary]==3.2.9
httpx==0.27.0
orjson==3.10.6
prometheus-client>=0.16.0
PyJWT==2.9.0

# --- Test dependencies ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
respx==0.20.2
